Destinations++
added unique menuItem sprite for ship destinations.
vaguely looks like a treasure map.

Item Costs++

if an item costs 0, then it's cost shouldn't be displayed
on any vendor screen. done. this makes it easier to
see destination menuItems, plus it conveys the 'cost'
of the item visually - there is no cost for this item.
this affect pets, as pets are all free. tested, working.