Pier Objs
pier objs connect to the boat's bridge.
this is reserved for civilized locations, where
a pier might be present in a maintained port.
in non-civilized locations, the swamp entrance
for example, there would be no such bridge to
a pier. instead link would just dive into the
water from the boat. done.

added animated water waves to the pier obj corners.