Map Screen++
there is an issue with map screen not resetting it's
scroll sprites when it's opened for the 2nd time..

created ResetScroll() in functions_scroll.
this is called from the Open() method of level map screen.
this ensures the scroll is reset to an opening state,
so when the screen tries to open the scroll, it actually
can... done, tested, working.

also improved the scroll sprites too, cause why not?