Prep for XML to CS rewrite
removed a variety of levels in order to simplify the
task of rewwriting the XML to CS pipeline to accomodate
for more islands.

created saveData id enum, to tag level data to island
biome (for level saving purposes).

editor is disabled by this push, due to how widgets
are looped thru to be displayed. I need to consolidate
textures soon anyway, and this will greatly affect how
widgts are displayed and handled, so this disabling
doesn't bother me too much right now.