XML to CS rewrite
instead of having all level data in a single .cs file,
there will be many .cs files specific for each island.
in order for this to be done well, the xml to cs pipeline
must change and should be made more modular to work on
island level sets. this is done using the islandID.

the islandID is an enum that becomes a global static public
class in the game. these global classes are then accessed by
the game upon the next build. this is how the level data is
stored inside the game, in binary format. this allows the game
to be easily ported to many different systems, including UWP.
however, .cs files can't get to big or be too complex, so
we have to break them into manageable sizes for visual studio.

created writeXMLtoCS() in DirectX backend class.
this handles lists of xml room data in conjunction with an
IslandID parameter to properly write each island set c# file.
for now, the boat level is being used as dummy data for each
island set. working as expected.

updated UWP project to ref new level data files.
actually playtested the UWP version. hmm. working.