Base Levels
to speed up development of certain levels, it would
be advantageous to create template levels that can be
copied and extended.

Base Boat + Sea Level

the boat can only be created in one position,
where link spawns. so, just create a base sea
level with boat, then use that level to build
more complex levels that depend upon the boat
+ sea objs / context. this also ensures that
the boat is built the same way each time, and
give the editor an idea of what level space is
available.

actually, this is not entirely true. the boat can
also be placed at a position where it is 'halfway'
in the level, still with stair access for later.
this gives the level more space for objs, and is
used in death mountains dungeon entrance level.

so, both base levels were created. then additional
planning was done based on the idea of base levels.