Editor++
the last display set from editor screen should display when
the screen is reopened(). it's a pita to have to click back
to the widget display set you need, just for one obj, over
and over again.

this was actually majorly easy to do: moved the forest state
initialization routine into the constructor for editor menu screen.
it was in the Open() method, which was resetting the widgets.
this change saves me MASSIVE amounts of time editing levels.