Destruction++
updated Functions_GameObject.BlowUp(RoomObj, Projectile)
with additional checks for all post objects. this means all
posts can now be destroy by anything in game that calls BlowUp(),
like bomb explosions, fireball explosions, lightning, etc..
this was easily done because the post is a blocking object,
and the codepath for projectile vs blocking object destruction
is very well established.

Thoughts

*there are issues with how fire spreads from explosions + posts.
basically, when a post is destroyed, it should spawn a ground fire
at it's location. this means the posts should exist outside of the
standard BlowUp() method. likely as fun_worldObj.BlowUpPost(Obj).

Thoughts

we should be modeling 'burnable' state as a boolean in objs.
this makes the check to add ground fires to objs much simpler,
and allows it to scale, as we are likely going to add more
burnable objects into the game. plus, we could likely consolidate
or simplify the concept of 'objects burning' in the codebase.
added Boolean burnable to GameObject class.