Burning GameObjects
in order to have a clear path forward for burning (and destroying)
gameobjs, we need to define and design the system to handle scale
and modifications / crazy ideas. this will be based on the
obj.burnable boolean added in the previous commit. also, that
previous commit wont compile. so.. yeah.

Burning - The Path Forward (psuedo code)

if obj.burnable, -> spawn ground fire at base of obj
FuncName: Functions_GameObject_World.BurnObject();
-this function spawns the ground fire, and kills() obj.

if !obj.burnable -> pass to established Fun_Obj.BlowUp();
-that's it, boss.

we just need to start flipping the burnable booleans in Obj.SetType().

Contention

the current system is setup to handle this by obj.type,
and this allow for a greater degree of control than by boolean
modeling. this scales linearly with obj definitions and intractions,
where modeling burnable would be less than linear, but give up
the ability to allow for unique burning interactions - for example
setting a torch on fire by having grass burn up to it. which, now
that i think about it - this should be possible within the game!
and we wouldn't be able to model the next layer of interaction
using a boolean 'burable' based approach. these decisions are so
small and subtle, but have so many consequences that are often
difficult to see quickly - or sometimes at all. leading to an
investment of work and a subsequent desire to believe 'that was
the right decision to make', when it was actually wrong - and
that takes time to realize. very glad i didn't commit to this
idea without running it through the best IDE there is: the brain.

This Commit

posts were given their own method for destruction, where
they will always spawn a ground fire. this allows for proper
spreading of fire between posts, keeping them consisten in
interaction with grass, bushes, etc.. 'burnable' objects.
this was also added into the region 'World Obj - special cases',
which was created specifically for object explosion interactions
such as these.