Flags++
place a bunch on flags on skull coliseum, port town houses,
and houses in the water. this is the start of something fun.
fixed left house's flag post to line up better with flag particle.
added many flag poles to skull island overworld pixels.

Bugs

fixed a really big bug related to loading levels that almost
brought me to tears. i'm really tired. i've been coding for
about 10-11 hours when this bug appears, and i just want to
get through it and go to sleep. it takes an hour to two hours
to do this. it was a nightmare. but i didn't give up.

playtested, working very well. goodnight.