Boss Door Bug
big bug: falling into a pit returns hero to last door
passed thru - this does not work on boss rooms, and
instead returns hero to the previous room's last door.
and this breaks the boss room and can put link into
a wall, locking him, breaking the game. fix this.

how we fix this: when hero enters a boss room,
this LevelSet.spawnPos_Dungeon XY is set to the room's only door.
this means any call to spawn in current room() puts link
at the bosses entrance door, and that kicks him forward into
the room again. learn more about the bug..

this door that hero touches and sets pos with,
is NOT A DOOR ROOMOBJECT. it is a piece of data
stored along with the rooms recs that define a dungeon.
do we have a 'door rec' on the boss room door? probably not.
because if we did, link would collide with it and the
code that you wrote, thats been tested, that has worked
until this point, is behaving as expected and you
probably just messed up. let's see..

yep. the logic to create the boss door is super wonky.
we gotta clean it up. here's what's happening: upon
dungeon generation, when we connect all the rooms
with door data pieces, we were skipping the door that
connected the boss room to the random room in the
cave dungeon, because the door connection routine
was only connecting boss rooms to hub rooms, with
door data. the fix is simple: the boss room always
has to connect to the hub room.

and this must be added when the hub room is created.
done and done. modified forest to be size 1, cave to
be size 2, swamp to be size 3. this is to test the
dungeon recipes tho. and to see if the map scroll
can display a size 3 dungeon properly.

lets test these dungeons out now.
added all dungeon entrances to the shadow king level.
compiled all the level data to cs, for easy testing.

upon testing, this bug is not fixed. but, the door
is indeed there. let's manually set the spawnPos for
link when he enters the boss room. this will be just
north of the entrance door..

done. tested and fixed.