Floor Blood++
fixed a bug where floor blood's animFrame was wrong.
as a result, dungeons are much messier (with blood).
however, I removed all hue from the blood, so it appears
more as a stain. this is for color harmony purposes,
until a better color for blood is chosen.

Dungeon Sheets++

updated dungeon textures to their color versions,
this is to ensure the texture loading routines are
working properly.

Exiting Dungeons++

pressing A (hero interaction) with an exit obj
should exit the dungeon. this allows a clipping
hero to exit the dungeon. otherwise, we have to
drop out of clipping, and that wastes seconds
we could be doing something else. plus, we jump
right back into clipping post-exit. so.... duh.
done. this makes testing much easier/faster.

game dungeons tested, working properly, tho
they could use *alot* of coloring love.