End Game Update to LSN Screen.
when player loads or saves or creates a game, the load/save/new
screen appears. now, when savedata shows all three dungeons have
been beaten, the text - "beaten!" appears next to the game name.
this is just a simple way to let the player know they did indeed
beat the game, as there is no end title or credits screen yet,
and i'm not in a big rush to make one.

implemented and tested, via xml editing.