Coliseum Taunting
*we should have a 'you have died shamefully in the shadow coliseum'
'the shadow king is entertained by your defeat.'

judge should check for hero's death and pop dialog if so.
this should only happen once, btw... done. this was implemented
using a semi-dirty hack: by checking the screens.dialog.dialogs
list against the global dialogs list to see if the failure dialog
had been set yet. if not, then initial entry point for failure dialog.
this sets the failure dialog, kicks it to player, then (due to the
logic check) exits the dialog creation section, waiting for the
final frame of the player's death before exiting the level (via
closing the screen, which slowly fades out the world to black).
tested.