Cleaning Up
updated design.md with words discussing development
of overworld. cleaned up ideas and setup a more realistic
roadmap, based on how I see various systems and design
choices coming together. I think items and interactions
need to improve alongside dungeons and locations, which
rely on underlying systems that really need to be updated.
so, that's what I'll handle in the next version, which I
think will improve the game in the 'right' direction.