Loading Saving Resetting GameData++
purchased bow, left level, bow was gone.

we loaded player data, which overwrote the values we
had changed in ram. this happened when we exited a level
or loaded the overworld level - prolly from overworld
load because we assumed we were coming from a lsn
screen (which isn't wrong, but it's wiping progress).

so, let's check the overworld load, and make it's coming
from that path too.. it's not happening in the overworld
open(), in base or shadowking class. def not in overworld
screen, but we cleaned up the class a little bit too.

moving onto functions_level and level screen..
right off the bat i see a call to func_hero.setLoadout()
in level screen, which was a call to make sure the
loadout was properly setup prior to play for an instance
based system design. this is only called by screel level,
when it opens. this was moved into the exit routine of
the load save new screen, so anytime that we leave that
screen, the player's type and loadout is sync'd to the
current data.

this fixes the bug where we lose items in 0.77.

however, testing reveals that we aren't properly loading
the player's last selected items from saved data.
we may of not saved this data, or we may default to
a standard loadout and not even try to load last loadout.
hmmm.... no, we have values for the current loadout
in savedata...

moved routines to set hero's loadout and type, and last
location into the close() method of lsn screen. this assumes
the current data is CURRENT (it is), and updates in game
values. this solves the loadout bug, AND the overworld
spawn bug - many tests were performed to ensure this works.