Bugs--
upon dying, loadout is reset. dont do that.
this was actually fixed as a consequnce of
the last commit, so we dont need to do this.
tested and working. however,

the summary screen works the first time.
the second time - not so much.

is this the screen or the dungoen record?

first, we should only display the summary
screen if link dies in a dungeon. if level
set.current points to the dungeon, then pop
summary screen, otherwise, we just kick hero
to the overworld screen via levelClose() and
exit action to overworld map.

second, we need to check that we are reset()ing
the dungeon record dr.reset(), just as we finish
building the dungeon - we do!

test that summary screen pops and works:
win/lose in dungeon - working. lose in field,
back to overworld we go!

exit action for level screen was updated to stop
the dungeon timer if the exit action was to summary
screen.

The Cause

if link uses a fairy to save himself from death,
then it clears to an unknown item.. which is why his
item disappears upon death in a tense battle (when
you're not looking at the item swap happening).
otherwise, the inventory and loadout systems are
working as designed. so upon calling HandleDeath()
for the hero, we simply store the hero's item in a
temp value, then restore that value post fairy
bottle routines. this fixes the annoying bug
of 'losing' your item upon a fairy-rez (self-rez).