Testing Enemy Actors using Player1 Input
we need to be able to turn hero into an enemy.
then, we can test the shooting routines, which test
the projectile spawn routines, which we are going to update
to use the actor's hitbox, instead of assuming they are
all just the hero, with the hero's specific hitbox.

this is done via pressing the f6 button in the editor,
as noted by the top menu now. this simply sets the hero's
type, using the actor.transform() method. this way, we
iterate through all enemy types, checking their animations
and interactions with the world. we can even load levels
up, then play with them as different enemy types.

THIS IS INCREDIBLY FUN.

now that this is setup and ready, this can be the foundation
for a number of dreams and mini-games, simply because we can
load a level, then change the hero's type to whatever enemy
type we want, and then if he leaves the field, we simply 'wake up'
or end the minigame, depending on how we want to do this.
very powerful and fun stuff here.