Miniboss OctoMouth
miniboss watery mouth can't throw fireballs. no item equipped.
fixed, magic fireball equipped, ai rewritten to call useItem.

but the hero can't use magic in the water - unless he is of
a specific enemy type - octomouth and kraken and tentacle.

non-hero actors can now use items in the water (while swimming).
this allows enemies in the water to use items, if they want to.

if the player is an actor type other than blob and link,
then they attack with their weapon instead of diving underwater.
a consequence, is that spiders now walk on water, and can
attack on water, so they are insanely scary. and an unarmored
spider moves very fast, and link would have no defense against
this enemy in the water, so this matchup is very overpowered.

modified octomouth's ai to dash when a direction is input,
so this should make him more active when he's underwater.

item use in water was modified to be based on the hero's
compMove.grounded boolean, which models if an actor is flying
or not. if the hero is a flying actor, then allow item use
while swimming, because they're actually hovering above the water.

playtested all enemies on land and on water. fun.
the unarmored spider is not as fast as i thought, because
we set the hero's friction to very slow when he goes
into a swim state. the spider is in a swim state, but the
swim anim frames redirect back to normal idle/walking frames.
so the spider appears to walk on water, but is slow. i like it.

update: link and the blob are only actors that cannot use items
in water now. hero, as any other actor type, can use items in water.

tested, most actors working as expected now.