Bush Growth
when a bush regrows (when it's near a water tile or
ditch filled with water), it becomes a blocking object.
it transforms from a non-blocking stump into a blocking
bush (which can be destroyed easily, but still blocking).
this is important because, one consequence of this system
is that a bush can regrow on the hero, locking him 'into'
the bush, making him unable to move. we handled that with
a specific check in bush regrowth routines - but now i
realize that this should be extended to include all actors,
simply because a bush can grow-lock minibosses or bosses,
like they did when i was playtesting 0.77 in the miniboss
room. in this room, we have a miniboss that shoots fireballs
in a room filled with water and bushes. this leads to the
cycle where fireballs burn bushes into stumps, which regrow
into bushes again. however, the bushes were grow locking
the miniboss, who would then still shoot fireballs, which
caused the fireball to immediately hit the bush, creating
an explosion, hurting the miniboss, causing him to go underwater,
and right as he popped back up - the bush would regrow -
continuing this cycle of never ending torture for the miniboss.
while this was entertaining to watch, this was not the
intended behavior and this commit prevents this state from
being possible - simply by preventing bush regrowth if
the stump touches any actor at all. which loops over all
actors, doing a collision check, which is mildly expensive.
oh well.