Water Logic and Actors ++
we want to preserve diving for specific actor types.
this is ocotmouth, kraken, tentacle. done. we can still
dive using these actor types. now we can further improve
them.

octomouth dives and shoots fireballs with player input. yep.
tentacle dives and attacks with use action, using fang. not yet.

tentacle should have fang as item, should use item.
when this happens, tentacle attack animation doesn't play.
this is for ai and actor input, so code isn't properly calling
the use item animation for tentacle..

this was caused by setAnimGroup(), which didn't set the use
animation group to attack anim, for any actor in a swimming
state. this was remedied using the same approach for allowing
attacking in water: we check the actor.underwaterEnemy boolean,
and allow underwater actors to attack and use items and weapons
like normal.

changed tentacle's ai to only set input, and not call
useItem, which caused a state unlock, allowing spamming
of attack/use while in water. this also fixed an invisible
bug with octomouth, and a variety of other enemies, which
would set use or attack to true, then call the useItem()
method, which eval'd(), but then the input to update routines
mapped input and caused the useItem() method to eval again,
essentially creating two projectiles everytime certain
enemies attacked. that's a no-no.

Bosses++

upgraded big eye to rarely spawn seekers while in phase 2,
which is chasing the hero.