Interaction Limiting
as hero, we can only begin a climb as link or blob.
only hero can grab while falling, and only as link or
blob. so both enttry points to climbing are closed to
non-link or non-blob actor types.

picking up, carrying, and throwing now limited to link
and blob. pushing or dragging objects now limited to
link and blob. updated pit interactions to pull all
actors in. this means some dungeon rooms containing
blobs will easily kill all the blobs within them, due
to the pits and the dumb bombs just charging at the
hero. which, is fine for now...