Throwing Objects++
the drawHeldObj() method is very inefficient.
the design of how we pickup & carry objects, is very inefficient.

what is the design? we model the picked up roomobj as a gameobj
reference to the roomObj. this is fine, if we wanted to put the
object down, as we did previously using B, but now we dash using
B and we also only throw the object and THIS IS IMPORTANT - the
thrown obj always explodes after becoming a power level 1 generic
projectile. in order to fulfill this design spec, we should be
modeling the heldObj as a projectile.

we are doing this using two refs in fun_hero.
carriedObj tracks to hero once it dupes the texture and animFrames
of the roomObj hero picks up. thrownObj inherits the carriedObjs
texture and animFrames, and handles traveling through the air
colliding and interacting with things. this makes logic simple,
plus it allows hero to pickup and carry an obj, while the thrown
projectile obj is still in play.

tested, all carry-able objects are also throw-able.
this includes the living roomObjs: crabs, rats, turtles.
modified pickup() to release the roomObj it dupes.
all thrown objs collide and interact as expected.

Bugs--
fixed a bug where bite wasn't dealing damage to actors