Prefix Removal
we need to drop 'projectile' prefix from projectilType enum values.
it's very redundant. done.

Throwing Enemies++

now when enemies are thrown, they create impact dust. this is different
from roomobjects and bushes, which create debris and leafs respectively.
this is just to remind myself that this can be different and should be
bloody later on.