Bugs--
enabling clipping freezes movement of hero. all input. except
start button. plus, all animation freezes too. in the entire game.
it's really weird. bug found: this was part of functions_pool.Update().
when doing interactions for the actor pool, we check to see if the actor
is the hero, and if the clipping boolean is enabled. in the scope we
then return, FROM THE ENTIRE POOL.UPDATE() method. whoops. instead of
returning, we should simply do nothing, because the else branch evaluates
roomObjs and projectiles against the actor, which we just want to prevent.
so - silly logical error by me that arose from migrating to the new
projectile systems. my bad. tested and working again.