Title++
upon title reload, the widget/window is already
open. it should reset itself. done.

added a sandbox enum value and menuitem to title,
which boots the game in sandbox mode with cheats
enabled and all items unlocked, for playtesting.

this required some more methods to be made for
flags and player data to be reset properly.