Spellbook++
when we close the spellbook, the inventory screen
should update the info widget to get the currently
selected items info - otherwise it displays the last
selected spell. done.

all inventory widgets should scale their menunItems
down in their Update(). always scale the menuItems down.
done. but we get double scaledown on selected item
on inventory screen. fixed.

no popup on spells screen, add that to move routine.
done. spells now scale up upon selection via A.

the spellbook should be equipped to the item slot,
when the wand is equipped to the weapon slot. done.
this allows for fast changing of spells, while
still letting player overwrite their current item
after equipping the wand.

if we equip the spell book to the item slot, the spell
book screen should open upon use. done.

*if we open the spellbook from the inventory screen,
we are selecting the spellbook item. don't reset the
inv's item to boomerang, just update the info widget.

spells screen hasn't reset and isn't updating the
info widget, so it's not opening to display anything.
BUT, we are drawing it.. so that's cool. fixed.

updated info widget to also match the speed of the
displayed menuitem sprite component.