Magic Systems++
functions_magic should have an update() method, that is called
from levelscreen.update(). this will allow magic to be called
sequentially over frames. we just have to make sure to stop
any magic that was casted or being casted across frames when we
exit the level screen (on a close screen routine). done.

we update magic systems, then the main pool updates. so,
any projectile magic creates will be collision and interaction
checked THAT FRAME.

Bombos++

bombos is the first spell to use (and test) the new magic systems.
now, explosions spawn left to right across the screen, projecting
down to world space, checking to see if they overlap the hero, and
then spawning explosions at these positions. this is done across
frames, in a controllable and timeable manner. tested and working,
i destroyed the town and it was insanely fun to watch.