Chain Explosions
this is a chain of explosions coming from caster's pos,
sequentially across frames. this is easy to do as a projectile.

Explosion Emitter Projectile

a projectile that simply creates explosions as it travels
around, for a life time. once dead, the explosions are no
longer emitted, and thus the effect ends.

explosive emitter added to game. tested, working.
this is really fun.