Magic++
improved chain explosions to be faster.
upgraded explosions to leave behind ground fires, making
them essesntially explosive and fire spells in one.
so because of that, fire and explosions have been merged
into explosive spells, and summons has come in for the
6th category instead of fire. upgraded Cast() to resolve
casters direction to cardinal, handling a case where caster
could be in a diagonal state, and the cast() wouldn't
complete due to only cardinal directions calling proper
routines.