Indestructible Object Widgets
we are adding the first ind widget. get this working with the
widget transitions in editor. then we grow it. working.
added 6 ind widgets, with temp dark block ind object.
selection working, but we need to rewrite how objs are
added/managed in editor before we can add indObjs to the
indestructible pool and actually start testing that system.
moving on..